+++ doc/Changelog.md
@@ -21,6 +21,7 @@ og:description: See what's new in the latest release of Roseau Load Flow !
 
 ## Unreleased
 
+- {gh-pr}`474` Wrap long element IDs in the hover box of the voltage profile plots.
 - {gh-pr}`473` Add support for custom color per element in the interactive map plots by accepting a callback as the
   `style_color` argument. The callable accepts two positional arguments: the element type and the element ID, and
   returns a color string or `None` to use the default color for that element.